Правки синтаксиса

+++ redmine_backup.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-. conf.sh
+source conf.sh
 
 use_message ()
 {
@@ -15,7 +15,7 @@ use_message ()
 mysql_backup ()
 {
 	echo "Snapshot backuping Redmine's MySQL db into Redmine instance..."
-	mysqldump --user=${REDMINE_DB_USER} --password=${REDMINE_DB_PASS} ${REDMINE_DB_NAME} > ${BACKUP_PATH}/${REDMINE_DB_BACKUP}
+	mysqldump --user="${REDMINE_DB_USER}" --password="${REDMINE_DB_PASS}" "${REDMINE_DB_NAME}" > "${BACKUP_PATH}"/"${REDMINE_DB_BACKUP}"
 	echo "($REDMINE_DB_BACKUP) done."
 	echo
 }
@@ -24,26 +24,26 @@ backup ()
 {
 	TYPE=$1
 	echo "Backuping Redmine's directory before sending it to a remote place..."
-	mkdir -pv ${BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}
-	cd ${BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE} || exit 1
-    mv ${BACKUP_PATH}/${REDMINE_DB_BACKUP} ${BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}
-	tar -czf ${REDMINE_BACKUP_NAME} ${REDMINE_HOME}
-	#tar -ztf ${REDMINE_BACKUP_NAME}
-	sha256sum ${REDMINE_BACKUP_NAME} > ${REDMINE_BACKUP_SHASUM_NAME}
-	tar -czf ../${REDMINE_BACKUP_DIR_NAME_}${TYPE}.tgz redmine*
-	cd ${BACKUP_PATH} || exit 1
-	${SCP} ${REDMINE_BACKUP_DIR_NAME_}${TYPE}.tgz ${SSH_USR}@${RHOST}:${REMOTE_BACKUP_PATH}
-    ${SCP} ${SCRIPT_DIR}/current_sum.sh ${SSH_USR}@${RHOST}:${REMOTE_BACKUP_PATH}
-    ${SSH} ${SSH_USR}@${RHOST} "bash ${REMOTE_BACKUP_PATH}/current_sum.sh"
+	mkdir -pv "${BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}" || exit 1
+	cd "${BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}" || exit 1
+    mv "${BACKUP_PATH}/${REDMINE_DB_BACKUP}" "${BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}"
+	tar -czf "${REDMINE_BACKUP_NAME}" "${REDMINE_HOME}" || exit 1
+	#tar -ztf "${REDMINE_BACKUP_NAME}" || exit 1
+	sha256sum "${REDMINE_BACKUP_NAME}" > "${REDMINE_BACKUP_SHASUM_NAME}" || exit 1
+	tar -czf ../"${REDMINE_BACKUP_DIR_NAME_}${TYPE}".tgz redmine* || exit 1
+	cd "${BACKUP_PATH}" || exit 1
+	${SCP} "${REDMINE_BACKUP_DIR_NAME_}${TYPE}".tgz "${SSH_USR}"@"${RHOST}":"${REMOTE_BACKUP_PATH}"
+    ${SCP} "${SCRIPT_DIR}"/current_sum.sh "${SSH_USR}"@"${RHOST}":"${REMOTE_BACKUP_PATH}"
+    ${SSH} "${SSH_USR}"@"${RHOST}" "bash ${REMOTE_BACKUP_PATH}/current_sum.sh"
 	if [[ $? -eq 0 ]]; then
-        	${SSH} ${SSH_USR}@${RHOST} "rm ${REMOTE_BACKUP_PATH}/current_sum.sh && rm ${REMOTE_BACKUP_PATH}/redmine{.tar.*,_mysql_*}"
-        	rm -rf ${BACKUP_PATH}/*
+        	${SSH} "${SSH_USR}"@"${RHOST}" "rm ${REMOTE_BACKUP_PATH}/current_sum.sh && rm ${REMOTE_BACKUP_PATH}/redmine{.tar.*,_mysql_*}"
+        	rm -rf "${BACKUP_PATH:?}/"*
         	echo "Check current archive chechsum done"
 	else
         echo "Check current archive chechsum filed"
 		while [[ $COUNT -ne 3 ]]; do
-			${SSH} ${SSH_USR}@${RHOST} "rm ${REMOTE_BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}.tgz"
-			rm -rf ${REDMINE_BACKUP_DIR_NAME_}${TYPE}
+			${SSH} "${SSH_USR}"@"${RHOST}" "rm ${REMOTE_BACKUP_PATH}/${REDMINE_BACKUP_DIR_NAME_}${TYPE}.tgz"
+			rm -rf "${REDMINE_BACKUP_DIR_NAME_}${TYPE}"
 			COUNT=$($COUNT + 1)
 			backup
 		done
@@ -64,13 +64,13 @@ remove_old ()
 	fi
 	echo "${MODEL_BKP_}${TYPE}.tgz"
 	echo "find ${REMOTE_BACKUP_PATH} -regex ${MODEL_BKP_}${TYPE}.tgz -mtime +${TIME} -exec rm {} \;"
-	${SSH} ${SSH_USR}@${RHOST} "find ${REMOTE_BACKUP_PATH} -regex ${MODEL_BKP_}${TYPE}.tgz -ctime +${TIME} -exec rm {} \;"
+	${SSH} "${SSH_USR}"@"${RHOST}" "find ${REMOTE_BACKUP_PATH} -regex ${MODEL_BKP_}${TYPE}.tgz -ctime +${TIME} -exec rm {} \;"
 }
 
 check_ssh ()
 {
     echo "Check SSH"
-	${SSH} ${SSH_USR}@${RHOST} "touch ${REMOTE_BACKUP_PATH}/${SSH_TEST_FILE} && rm -f ${REMOTE_BACKUP_PATH}/${SSH_TEST_FILE}"
+	${SSH} "${SSH_USR}"@"${RHOST}" "touch ${REMOTE_BACKUP_PATH}/${SSH_TEST_FILE} && rm -f ${REMOTE_BACKUP_PATH}/${SSH_TEST_FILE}"
 	if [[ $? != 0 ]]; then
 		echo " Can not access to remote server !"
 		echo " Check: hostname, remore user, ssh-keys, directory permissions !"
@@ -82,10 +82,10 @@ fi
 check_remote_sums ()
 {
     echo "Check remote sums"
-    ${SCP} checksum.sh ${SSH_USR}@${RHOST}:${REMOTE_BACKUP_PATH}
-	${SSH} ${SSH_USR}@${RHOST} "bash ${REMOTE_BACKUP_PATH}/checksum.sh"
+    ${SCP} checksum.sh "${SSH_USR}"@"${RHOST}":"${REMOTE_BACKUP_PATH}"
+	${SSH} "${SSH_USR}"@"${RHOST}" "bash ${REMOTE_BACKUP_PATH}/checksum.sh"
     if [[ $? -eq 0 ]]; then
-        ${SSH} ${SSH_USR}@${RHOST} "rm ${REMOTE_BACKUP_PATH}/checksum.sh && rm ${REMOTE_BACKUP_PATH}/redmine{.tar.*,_mysql_*}"
+        ${SSH} "${SSH_USR}"@"${RHOST}" "rm ${REMOTE_BACKUP_PATH}/checksum.sh && rm ${REMOTE_BACKUP_PATH}/redmine{.tar.*,_mysql_*}"
         echo "Check remote sums done"
     else
         echo "Check remote sums failed."
